% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Title is itititit}
\usage{
foo(n)
}
\arguments{
\item{n}{number of rows we have in the resulting first element data frame.}
}
\value{
A lst with two elements, the first is ... the second is ...
}
\description{
This is just a dummy functon.
}
\examples{
foo(2)
foo(10)
}
