% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoul2-package.R
\docType{package}
\name{seoul2-package}
\alias{seoul2}
\alias{seoul2-package}
\title{seoul2: Seoul Lecture}
\description{
Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shstat1729.github.io/seoul2/}
}

}
\author{
\strong{Maintainer}: Sang Ho Park \email{shstat1729@gmail.com}

}
\keyword{internal}
